Failed Attempts Convert Tuples to Dict, DF success
It shows various ways to try to convert some tuples to a dictionary but all the native Python suggestions tried didn't work on Python 3.12.  Pandas DataFrame easily succeeded.  Dictionary is desired to present a .json file on an API.
I will post to Stackoverflow when I figure out how to make it accept my output as text, not Python3.